refactor :
1. 문자열 ingredient 필드 삭제
2. volume은 실수값으로 받도록 수정

+++ admin/src/main/java/org/cocktail/admin/domain/cocktailingredient/converter/CockTailIngredientConverter.java
@@ -19,7 +19,7 @@ public class CockTailIngredientConverter {
 
         return CocktailIngredientEntity.builder()
                 .ingredient(byName)
-                .volume(Integer.valueOf(cocktailIngredientRequest.getVolume()))
+                .volume(Double.valueOf(cocktailIngredientRequest.getVolume()))
                 .unit(Unit.valueOf(cocktailIngredientRequest.getUnit()))
                 .build();
     }
